#58 fix: homeusecase 삭제 및 EggUseCase 분리
