+++ Walkie-iOS/Walkie-iOS/Sources/Data/Repository/Egg/EggRepository.swift
@@ -11,5 +11,5 @@ protocol EggRepository {
     func getEggsList(dummy: Bool) -> AnyPublisher<[EggEntity], NetworkError>
     func getEggDetail(dummy: Bool, eggId: Int) -> AnyPublisher<EggDetailEntity, NetworkError>
     func patchEggStep(requestBody: PatchEggStepRequestDto) -> AnyPublisher<Void, NetworkError>
-    func getEggsCount() -> AnyPublisher<EggsCountEntity, NetworkError>
+    func getEggsCount() -> AnyPublisher<Int, NetworkError>
 }
