+++ Walkie-iOS/Walkie-iOS/Sources/Domain/UseCase/Egg/Implementation/DefaultGetEggCountUseCase.swift
@@ -1,0 +1,16 @@
+//
+//  DefaultGetEggCountUseCase.swift
+//  Walkie-iOS
+//
+//  Created by ahra on 4/2/25.
+//
+
+import Combine
+
+final class DefaultGetEggCountUseCase: BaseEggUseCase, GetEggCountUseCase {
+    
+    func getEggsCount() -> AnyPublisher<Int, NetworkError> {
+        eggRepository.getEggsCount()
+            .mapToNetworkError()
+    }
+}
