+++ Walkie-iOS/Walkie-iOS/Sources/Domain/UseCase/Egg/Protocol/GetEggCountUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  GetEggCountUseCase.swift
+//  Walkie-iOS
+//
+//  Created by ahra on 4/2/25.
+//
+
+import Combine
+
+protocol GetEggCountUseCase {
+    func getEggsCount() -> AnyPublisher<Int, NetworkError>
+}
