+++ Walkie-iOS/Walkie-iOS/Sources/Domain/UseCase/Egg/Protocol/GetEggPlayUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  GetEggPlayUseCase.swift
+//  Walkie-iOS
+//
+//  Created by ahra on 4/1/25.
+//
+
+import Combine
+
+protocol GetEggPlayUseCase {
+    func getEggPlaying() -> AnyPublisher<EggEntity, NetworkError>
+}
